replace variadic with slice
When database/sql delegates to the driver Query or Exec functions, there
will be only one argument, a slice. The slice is not expanded to
variadic arguments. This change makes it clear that what the query and
exec functions will be passed will be a single slice of args, and not a
a slice with a single element that is a slice.